# apps/backend/api-spec.yaml
openapi: 3.0.1
info:
  title: Agent Builder API
  description: Autonomous Software Factory - AI-powered application development platform
  version: 1.0.0
  contact:
    name: Agent Builder Team
    email: support@agent-builder.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.agent-builder.app/dev
    description: Development server
  - url: https://api.agent-builder.app/prod
    description: Production server
security:
  - CognitoAuth: []
paths:
...
  /projects/{id}:
    get:
      summary: Get project details
      description: Retrieve detailed information about a specific project
      tags:
        - Projects
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update project
      description: Update project details
      tags:
        - Projects
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectName:
                  type: string
                requestPrompt:
                  type: string
                status:
                  type: string
                  enum: [PENDING, IN_PROGRESS, COMPLETED, FAILED]
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete project
      description: Delete a project and all associated data
      tags:
        - Projects
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '204':
          description: Project deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    CognitoAuth:
      type: apiKey
      in: header
      name: Authorization
      description: AWS Cognito JWT token in the format "Bearer {token}"
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      description: Unique identifier for the project
      schema:
        type: string
        format: uuid
  schemas:
...
    Project:
      type: object
      required:
        - id
        - userId
        - projectName
        - requestPrompt
        - status
      properties:
        id:
          type: string
          description: Unique project identifier
        userId:
          type: string
          description: Owner user ID
        projectName:
          type: string
          description: Human-readable project name
        requestPrompt:
          type: string
          description: Original user request/prompt
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED, FAILED]
          description: Current project status
        createdAt:
          type: string
          format: date-time
          description: Project creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code or type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "VALIDATION_ERROR"
            message: "Missing required field: projectName"
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "UNAUTHORIZED"
            message: "Authentication required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "NOT_FOUND"
            message: "Resource not found"
...